+++ catsoop/base_context.py
@@ -296,8 +296,9 @@ Special: user interface configuration flags
 
 # user information data configuration flags
 
-cs_user_config = {'section_variable': 'section', # variable for class section
-                  'default_section_name': 'default',
+cs_user_config = {
+    "section_variable": "section",  # variable for class section
+    "default_section_name": "default",
 }
 """
 Special: user information data configuration flags (e.g. used for groups)